refactor demo project to be Kotlin/MP
