+++ demo/README.md
@@ -1,5 +1,5 @@
 # Parsus Demo
 
 This project contains multiple grammars written with Parsus:
-* [Arithmetic expression parser and calculator](src/main/kotlin/me/alllex/parsus/demo/Arithmetic.kt)
-* [Boolean expression parser](src/main/kotlin/me/alllex/parsus/demo/BooleanExpression.kt)
+* [Arithmetic expression parser and calculator](src/commonMain/kotlin/me/alllex/parsus/demo/Arithmetic.kt)
+* [Boolean expression parser](src/commonMain/kotlin/me/alllex/parsus/demo/BooleanExpression.kt)
