+++ demo/src/commonMain/kotlin/me/alllex/parsus/demo/Arithmetic.kt
@@ -1,6 +1,6 @@
-@file:Suppress("MemberVisibilityCanBePrivate")
+@file:Suppress("MemberVisibilityCanBePrivate", "PackageDirectoryMismatch")
 
-package me.alllex.parsus.demo
+package me.alllex.parsus.demo.arithmetic
 
 import me.alllex.parsus.parser.*
 import me.alllex.parsus.token.literalToken
