+++ demo/src/commonMain/kotlin/me/alllex/parsus/demo/BooleanExpression.kt
@@ -1,8 +1,8 @@
-@file:Suppress("MemberVisibilityCanBePrivate")
+@file:Suppress("MemberVisibilityCanBePrivate", "PackageDirectoryMismatch")
 
-package me.alllex.parsus.demo
+package me.alllex.parsus.demo.boolean
 
-import me.alllex.parsus.demo.BooleanExpression.*
+import me.alllex.parsus.demo.boolean.BooleanExpression.*
 import me.alllex.parsus.parser.*
 import me.alllex.parsus.token.literalToken
 import me.alllex.parsus.token.regexToken
